add static of all the funcs

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: xinwang <xinwang@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/17 16:57:30 by xinwang           #+#    #+#             */
-/*   Updated: 2019/10/17 20:35:23 by xinwang          ###   ########.fr       */
+/*   Updated: 2019/10/18 04:19:57 by xinwang          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ char        *ft_strdup(const char *s);
 char        *ft_strjoin(char const *s1, char const *s2);
 size_t		ft_strlcat(char *dst, const char *src, size_t size);
 size_t		ft_strlcpy(char *dest, const char *src, size_t size);
-int			ft_strlen(char *str);
+size_t			ft_strlen(char *str);
 char        *ft_strmapi(char const *s, char (*f)(unsigned int, char));
 int			ft_strncmp(const char *s1, const char *s2, size_t n);
 char      	*ft_strnstr(const char *str, const char *to_find, size_t len);
